+++ src/osgPlugins/flt/ReaderWriterATTR.cpp
@@ -49,8 +49,8 @@ typedef signed char     int8;
 typedef unsigned char   uint8;
 typedef signed short    int16;
 typedef unsigned short  uint16;
-typedef signed long     int32;
-typedef unsigned long   uint32;
+typedef signed int	int32;
+typedef unsigned int    uint32;
 typedef float           float32;
 typedef double          float64;
 
@@ -75,20 +75,20 @@ static void endian2(void* pSrc, int nSrc, void* pDst)
     }
     else if (nSrc == 4)
     {
-        long tmp1;
-        tmp1 = *(long *)pSrc;
+        uint32 tmp1;
+        tmp1 = *(uint32 *)pSrc;
         tmp1 = (tmp1 << 24) | ((tmp1 << 8) & 0xff0000) | ((tmp1 >> 8) & 0xff00) | ((tmp1 >> 24) & 0xff);
-        *(long *)pDst = tmp1;
+        *(uint32 *)pDst = tmp1;
     }
     else if (nSrc == 8)
     {
-        long tmp1, tmp2;
-        tmp1 = *(long *)pSrc;
-        tmp2 = *(1 + (long *)pSrc);
+        uint32 tmp1, tmp2;
+        tmp1 = *(uint32 *)pSrc;
+        tmp2 = *(1 + (uint32 *)pSrc);
         tmp1 = (tmp1 << 24) | ((tmp1 << 8) & 0xff0000) | ((tmp1 >> 8) & 0xff00) | ((tmp1 >> 24) & 0xff);
         tmp2 = (tmp2 << 24) | ((tmp2 << 8) & 0xff0000) | ((tmp2 >> 8) & 0xff00) | ((tmp2 >> 24) & 0xff);
-        *(long *)pDst = tmp2;
-        *(1 + (long *)pDst) = tmp1;
+        *(uint32 *)pDst = tmp2;
+        *(1 + (uint32 *)pDst) = tmp1;
     }
 }
